requirements.txt
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer
python-dotenv
streamlit
langchain
openai
requests
langchain-community
langgraph
langchain-openai
Flask
streamlit-option-menu
streamlit
requests
streamlit_extras
ipython
pyodbc
wikipedia